Refactored code and added python API
